Amended the test.
